+++ tests/conftest.py
@@ -7,3 +7,42 @@
 # @License: BSD 3-clause (http://www.opensource.org/licenses/BSD-3-Clause)
 
 from __future__ import annotations
+
+import asyncio
+
+import pytest
+from pytest_mock import MockerFixture
+
+from clu.testing import setup_test_actor
+
+from lvmopstools.actor import ActorState, ErrorCodes, LVMActor
+
+
+class TestActor(LVMActor):
+    async def _check_internal(self):
+        pass
+
+    async def _troubleshoot_internal(
+        self,
+        error_code: ErrorCodes,
+        exception: Exception | None = None,
+    ):
+        return True
+
+
+@pytest.fixture()
+async def lvm_actor(mocker: MockerFixture):
+    actor = TestActor(name="test_actor", check_interval=1)
+
+    await setup_test_actor(actor)  # type: ignore
+
+    actor.update_state(ActorState.READY)
+    actor._check_task = asyncio.create_task(actor._check_loop())
+
+    actor._check_internal = mocker.AsyncMock(return_value=None)
+    actor._troubleshoot_internal = mocker.AsyncMock(return_value=True)
+    actor.is_connected = mocker.MagicMock(return_value=True)
+
+    yield actor
+
+    await actor.stop()