+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Next version
+
+### 🚀 New
+
+* Added `LVMActor` class with actor state and troubleshooting framework.
+
 ## 0.1.0 - January 19, 2024
 
 ### 🚀 New